Moved class into translator.py
